src/sprint3_lambda/requirements.txt
psycopg2-binary==2.9.6
jmespath==1.0.1